+++ app/src/main/java/com/devwithzachary/completelinuxinstaller/ui/screens/about/AboutScreen.kt
@@ -312,6 +312,9 @@ fun AboutScreen() {
             }
         }
 
+        // Project Credits & Contributors Section
+        AboutCreditsSection()
+
         // Changelog History Section
         AboutChangelogSection()
     }